Moved Theme to ThemeManager
